+++ app/Http/Controllers/ApiController.php
@@ -1,0 +1,12 @@
+<?php namespace JobApis\JobsToMail\Http\Controllers;
+
+use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Routing\Controller as BaseController;
+use Illuminate\Foundation\Validation\ValidatesRequests;
+
+class ApiController extends BaseController
+{
+    use DispatchesJobs, ValidatesRequests;
+
+    // Coming soon.
+}