+++ app/Http/Requests/CreateUser.php
@@ -22,7 +22,10 @@ class CreateUser extends FormRequest
     public function messages()
     {
         return [
-            'regex' =>'The location should be formatted "City, ST". Currently only works for US locations.'
+            'regex' =>'The location should be formatted 
+                "City, ST". Currently only works for US locations.',
+            'premium' =>'Only premium users can use this feature. 
+                Contact upgrade@jobstomail.com to get access.',
         ];
     }
 
@@ -37,6 +40,7 @@ class CreateUser extends FormRequest
             'email' => 'required|email',
             'keyword' => 'required',
             'location' => 'required|regex:/([^,]+), \s*(\w{2})/',
+            'no_recruiters' => 'required|premium',
         ];
     }
 }