+++ resources/views/searches/index.blade.php
@@ -13,20 +13,7 @@
             <div class="card card-block">
                 <h4 class="card-title">"{{ $search->keyword }}"</h4>
                 <p>Location: "{{ $search->location }}"</p>
-                <div class="form-check">
-                    <label class="form-check-label disabled"
-                           data-toggle="tooltip"
-                           data-html="true"
-                           data-placement="left"
-                           title="Enable this option by contacting <a href='mailto:upgrade@jobstomail.com' target='_blank'>upgrade@jobstomail.com</a>">
-                        <input type="checkbox"
-                               class="form-check-input"
-                               {{ $search->no_recruiters ? 'checked' : '' }}
-                               disabled
-                               readonly>
-                        <span class="">Don't include posts by recruiters</span>
-                    </label>
-                </div>
+                @include('searches.components.no-recruiters-checkbox')
                 <p><a href="/searches/{{ $search->id }}/unsubscribe" class="card-link">Unsubscribe</a></p>
             </div>
         @endforeach